dev: refactored root components to a pages directory, and added remainint path aliases
